test: recreate migrations
