--- api/alembic/versions/75705c1cfbfa_create_posts_table.py
@@ -1,38 +1,0 @@
-"""create posts table
-
-Revision ID: 75705c1cfbfa
-Revises: 
-Create Date: 2024-06-10 04:35:13.561516
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '75705c1cfbfa'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('posts',
-    sa.Column('id', sa.UUID(), nullable=False),
-    sa.Column('title', sa.String(), nullable=False),
-    sa.Column('content', sa.String(), nullable=False),
-    sa.Column('is_published', sa.Boolean(), server_default='TRUE', nullable=False),
-    sa.Column('created_at', sa.TIMESTAMP(timezone=True), nullable=False),
-    sa.Column('updated_at', sa.TIMESTAMP(timezone=True), nullable=False),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('posts')
-    # ### end Alembic commands ###
