--- api/alembic/versions/c9c30513cd5b_create_votes_table.py
@@ -1,38 +1,0 @@
-"""create votes table
-
-Revision ID: c9c30513cd5b
-Revises: f879f6717f05
-Create Date: 2024-06-16 03:22:55.861155
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'c9c30513cd5b'
-down_revision: Union[str, None] = 'f879f6717f05'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('votes',
-    sa.Column('id', sa.Integer(), autoincrement=True, nullable=False),
-    sa.Column('post_id', sa.UUID(), nullable=False),
-    sa.Column('user_id', sa.UUID(), nullable=False),
-    sa.Column('created_at', sa.TIMESTAMP(timezone=True), nullable=False),
-    sa.ForeignKeyConstraint(['post_id'], ['posts.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table('votes')
-    # ### end Alembic commands ###
