--- api/alembic/versions/f879f6717f05_add_user_fk_to_posts_table.py
@@ -1,32 +1,0 @@
-"""add user fk to posts table
-
-Revision ID: f879f6717f05
-Revises: 3ac539e096fb
-Create Date: 2024-06-15 03:07:02.283567
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'f879f6717f05'
-down_revision: Union[str, None] = '3ac539e096fb'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('posts', sa.Column('user_id', sa.UUID(), nullable=False))
-    op.create_foreign_key(None, 'posts', 'users', ['user_id'], ['id'], ondelete='CASCADE')
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_constraint(None, 'posts', type_='foreignkey')
-    op.drop_column('posts', 'user_id')
-    # ### end Alembic commands ###
